test-requirements.txt
ansible-lint==5.4.0
cryptography
docker==5.0.0
flake8==3.9.2
jmespath>=0.9.4
molecule-docker==0.2.4
molecule==3.4.0
netaddr>=0.8.0
pytest-testinfra==6.4.0
pytest==6.2.5
tox-gh-actions==2.6.0
tox==3.24.3
yamllint==1.26.3
